[FIX] Fixing Code Duplication and Tests
