+++ src/components/AddUserOrgAdmin.js
@@ -1,25 +1,12 @@
-import React, { useState } from 'react';
-import AddUserForm from './AddUserForm'; 
+import React from 'react';
+import AddUserForm from './AddUserForm';
+import useAddUserLogic from './useAddUserLogic';
 
 const AddUserOrgAdmin = () => {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [email, setEmail] = useState('');
-  const [position, setPosition] = useState('');
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    // Implement logic to submit user data
-    console.log('Submitted user data:', { username, password, email, position });
-    // Clear form after submit
-    setUsername('');
-    setPassword('');
-    setEmail('');
-    setPosition('');
-  };
+  const { handleSubmit } = useAddUserLogic();
 
   return (
-    <AddUserForm title="Add User (Momofin Admin)" onSubmit={handleSubmit} />
+    <AddUserForm title="Add User (Org Admin)" onSubmit={handleSubmit} />
   );
 };
